requirements.txt
torch==1.4.0
torchvision==0.5.0
numpy==1.16.1
matplotlib
svgwrite
cairosvg
IPython
six>=1.12.0
tensorflow
tensorboardX
pillow
svglib
tqdm
jupyter
scikit-image
pandas
networkx
moviepy
numba
sklearn
umap-learn
umap-learn[plot]
shapely
kivy